+++ src/main/java/insane96mcp/survivalreimagined/module/hungerhealth/fooddrinks/CustomFoodProperties.java
@@ -3,21 +3,33 @@ package insane96mcp.survivalreimagined.module.hungerhealth.fooddrinks;
 import com.google.gson.*;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.reflect.TypeToken;
+import com.mojang.datafixers.util.Pair;
 import insane96mcp.insanelib.util.IdTagMatcher;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Supplier;
 
 @JsonAdapter(CustomFoodProperties.Serializer.class)
 public class CustomFoodProperties extends IdTagMatcher {
 	public int nutrition = -1;
 	public float saturationModifier = -1f;
 	public int eatingTime = -1;
-	public Boolean fastEating = null;
+	public boolean fastEating = false;
+	//If null, don't modify existing effects
+	public List<Pair<Supplier<MobEffectInstance>, Float>> effects = null;
 
 	public CustomFoodProperties(Type type, String id) {
 		super(type, id);
+	}
+
+	public List<Pair<MobEffectInstance, Float>> getEffects() {
+		return this.effects.stream().map(pair -> Pair.of(pair.getFirst() != null ? pair.getFirst().get() : null, pair.getSecond())).collect(java.util.stream.Collectors.toList());
 	}
 
 	public static class Builder {
@@ -50,11 +62,23 @@ public class CustomFoodProperties extends IdTagMatcher {
 			this.customFoodProperties.fastEating = true;
 			return this;
 		}
+
+		public Builder noEffects() {
+			this.customFoodProperties.effects = new ArrayList<>();
+			return this;
+		}
+
+		public Builder addEffect(MobEffectInstance mobEffectInstance, float chance) {
+			if (this.customFoodProperties.effects == null)
+				this.customFoodProperties.effects = new ArrayList<>();
+			this.customFoodProperties.effects.add(Pair.of(() -> mobEffectInstance, chance));
+			return this;
+		}
 	}
 
 	@Override
 	public String toString() {
-		return "FoodValue{type: %s, id: %s, nutrition: %s, saturation_modifier: %s, eating_time: %s, fast_eating: %s}".formatted(this.type, this.location, this.nutrition, this.saturationModifier, this.eatingTime, this.fastEating);
+		return "FoodValue{type: %s, id: %s, nutrition: %s, saturation_modifier: %s, eating_time: %s, fast_eating: %s, effects: %s}".formatted(this.type, this.location, this.nutrition, this.saturationModifier, this.eatingTime, this.fastEating, this.effects);
 	}
 
 	public static final java.lang.reflect.Type LIST_TYPE = new TypeToken<ArrayList<CustomFoodProperties>>(){}.getType();
@@ -65,21 +89,21 @@ public class CustomFoodProperties extends IdTagMatcher {
 			String id = GsonHelper.getAsString(json.getAsJsonObject(), "id", "");
 			String tag = GsonHelper.getAsString(json.getAsJsonObject(), "tag", "");
 
-			if (!id.equals("") && !ResourceLocation.isValidResourceLocation(id)) {
+			if (!id.isEmpty() && !ResourceLocation.isValidResourceLocation(id)) {
 				throw new JsonParseException("Invalid id: %s".formatted(id));
 			}
-			if (!tag.equals("") && !ResourceLocation.isValidResourceLocation(id)) {
+			if (!tag.isEmpty() && !ResourceLocation.isValidResourceLocation(id)) {
 				throw new JsonParseException("Invalid tag: %s".formatted(tag));
 			}
 
 			CustomFoodProperties customFoodProperties;
-			if (!id.equals("") && !tag.equals("")){
+			if (!id.isEmpty() && !tag.isEmpty()){
 				throw new JsonParseException("Invalid object containing both tag (%s) and id (%s)".formatted(tag, id));
 			}
-			else if (!id.equals("")) {
+			else if (!id.isEmpty()) {
 				customFoodProperties = new CustomFoodProperties(Type.ID, id);
 			}
-			else if (!tag.equals("")){
+			else if (!tag.isEmpty()){
 				customFoodProperties = new CustomFoodProperties(Type.TAG, tag);
 			}
 			else {
@@ -93,29 +117,55 @@ public class CustomFoodProperties extends IdTagMatcher {
 				customFoodProperties.fastEating = GsonHelper.getAsBoolean(json.getAsJsonObject(), "fast_eating");
 			}
 
+			if (json.getAsJsonObject().has("effects")) {
+				JsonArray effects = json.getAsJsonObject().getAsJsonArray("effects");
+				if (customFoodProperties.effects == null)
+					customFoodProperties.effects = new ArrayList<>();
+				effects.forEach(element -> {
+					String stringId = GsonHelper.getAsString(element.getAsJsonObject(), "id");
+					ResourceLocation effectId = ResourceLocation.tryParse(stringId);
+					MobEffect mobEffect = ForgeRegistries.MOB_EFFECTS.getValue(effectId);
+					if (mobEffect == null)
+						throw new JsonParseException("Mob effect %s not found".formatted(stringId));
+					int amplifier = GsonHelper.getAsInt(element.getAsJsonObject(), "amplifier", 0);
+					int duration = GsonHelper.getAsInt(element.getAsJsonObject(), "duration");
+					MobEffectInstance effectInstance = new MobEffectInstance(mobEffect, duration, amplifier);
+					float chance = GsonHelper.getAsFloat(element.getAsJsonObject(), "chance", 1f);
+					customFoodProperties.effects.add(Pair.of(() -> effectInstance, chance));
+				});
+			}
+
 			return customFoodProperties;
 		}
 
 		@Override
 		public JsonElement serialize(CustomFoodProperties src, java.lang.reflect.Type typeOfSrc, JsonSerializationContext context) {
 			JsonObject jsonObject = new JsonObject();
-			if (src.type == Type.ID) {
+			if (src.type == Type.ID)
 				jsonObject.addProperty("id", src.location.toString());
-			}
-			else if (src.type == Type.TAG) {
+			else if (src.type == Type.TAG)
 				jsonObject.addProperty("tag", src.location.toString());
-			}
-			if (src.nutrition >= 0) {
+
+			if (src.nutrition >= 0)
 				jsonObject.addProperty("nutrition", src.nutrition);
-			}
-			if (src.saturationModifier >= 0f) {
+			if (src.saturationModifier >= 0f)
 				jsonObject.addProperty("saturation_modifier", src.saturationModifier);
-			}
-			if (src.eatingTime >= 0) {
+			if (src.eatingTime >= 0)
 				jsonObject.addProperty("eating_time", src.eatingTime);
-			}
-			if (src.fastEating != null) {
-				jsonObject.addProperty("fast_eating", src.fastEating);
+			if (src.fastEating)
+				jsonObject.addProperty("fast_eating", true);
+			if (src.effects != null) {
+				JsonArray effects = new JsonArray();
+				src.getEffects().forEach(pair -> {
+					JsonObject effect = new JsonObject();
+					effect.addProperty("id", ForgeRegistries.MOB_EFFECTS.getKey(pair.getFirst().getEffect()).toString());
+					effect.addProperty("amplifier", pair.getFirst().getAmplifier());
+					effect.addProperty("duration", pair.getFirst().getDuration());
+					if (pair.getSecond() < 1f)
+						effect.addProperty("chance", pair.getSecond());
+					effects.add(effect);
+				});
+				jsonObject.add("effects", effects);
 			}
 
 			return jsonObject;